requirements.txt
datasets
pinecone
python-dotenv
openai
weaviate-client
langchain
langchain-community 
langchain-openai
langgraph
langchain_anthropic
firecrawl-py
